feat: refactored card hand, enemy card hand
