+++ Assets/Scenes/SampleScene.unity
@@ -272,14 +272,14 @@ MonoBehaviour:
   cardStack: {fileID: 259494340}
   stackPos: {fileID: 1876400315}
   cardSpawnMoveSpeed: 50
-  cardViewMoveSpeed: 15
   cardRotationSpeed: 50
   spacing: 1.5
   maxWidth: 500
   radius: 50
-  moveOverDistance: 5
   raisedAmount: 3
   fadeThreshold: 3
+  cardViewMoveSpeed: 15
+  moveOverDistance: 5
 --- !u!114 &259494340
 MonoBehaviour:
   m_ObjectHideFlags: 0
@@ -1534,15 +1534,15 @@ Transform:
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
   m_GameObject: {fileID: 703638986}
-  m_LocalRotation: {x: 0, y: 1, z: 0, w: 0}
-  m_LocalPosition: {x: 0, y: 6.82, z: 4.2}
+  m_LocalRotation: {x: -1, y: 0.0000002533197, z: -0, w: 0}
+  m_LocalPosition: {x: 0, y: 8.63, z: 4.2}
   m_LocalScale: {x: 1, y: 1, z: 1}
   m_ConstrainProportionsScale: 0
   m_Children:
   - {fileID: 324752108}
   m_Father: {fileID: 783440548}
   m_RootOrder: 4
-  m_LocalEulerAnglesHint: {x: 0, y: 180, z: 0}
+  m_LocalEulerAnglesHint: {x: 0, y: 180, z: -180}
 --- !u!114 &703638988
 MonoBehaviour:
   m_ObjectHideFlags: 0
@@ -1590,23 +1590,20 @@ MonoBehaviour:
   m_GameObject: {fileID: 703638986}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 300be5ea9a4e3fe47988d7e3296b33e9, type: 3}
+  m_Script: {fileID: 11500000, guid: ffb8c565b3c6db7489d81e3174f0d720, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
   uiCam: {fileID: 783440547}
-  cardPrefab: {fileID: 5226560222687344958, guid: 0b976eb92b7b2ab46b27abc708faa38e,
+  cardPrefab: {fileID: 5226560222687344958, guid: effb4baf131c236458a55125c76878ef,
     type: 3}
   cardStack: {fileID: 703638988}
   stackPos: {fileID: 1876400315}
   cardSpawnMoveSpeed: 50
-  cardViewMoveSpeed: 15
   cardRotationSpeed: 50
   spacing: 1.5
   maxWidth: 500
   radius: 50
-  moveOverDistance: 5
-  raisedAmount: 3
-  fadeThreshold: 3
+  raisedAmount: 5
 --- !u!1 &783440544
 GameObject:
   m_ObjectHideFlags: 0
