+++ Assets/Scripts/CardSystem/BaseCardBehaviour.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: f9e1c5287a0321a4b87b744134183921
+guid: eea9d070330db80478fe7241e316e19a
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
