+++ Assets/Scripts/CardSystem/EnemyCardBehaviour.cs
@@ -1,0 +1,3 @@
+﻿public class EnemyCardBehaviour : BaseCardBehaviour {
+
+}
