+++ Assets/Scripts/CardSystem/PlayerCardBehaviour.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: f9e1c5287a0321a4b87b744134183921
+guid: 64089e59f64bd0a4ba8f602c3a208a5c
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
